refactor(condo): DOMA-7169 added useEmployeeRoleTableData hook
